refactor: rework pull_request.synchronize event handler
