+++ src/config.ts
@@ -3,6 +3,7 @@ import StarCreateHandler from "./events/star.created";
 import ReleaseHandler from "./events/release.released";
 import PullRequestOpenHandler from "./events/pull_request.opened";
 import PullRequestCloseHandler from "./events/pull_request.closed";
+import PullRequestSynchronizeHandler from "./events/pull_request.syncronize";
 import IssueCloseHandler from "./events/issues.closed";
 import IssueOpenHandler from "./events/issues.opened";
 
@@ -15,6 +16,7 @@ export const Handlers: HandlerModule[] = [
   ReleaseHandler,
   PullRequestOpenHandler,
   PullRequestCloseHandler,
+  PullRequestSynchronizeHandler,
   IssueCloseHandler,
   IssueOpenHandler,
 ];
